pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "temporal-framework"
version = "0.1.0"
description = "Temporal Dimension Enhancement: 6-Tuple Contextual Integrity Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.2",
    "PyYAML>=6.0",
    "python-dateutil>=2.8.0",
    "neo4j>=5.0.0",
    "graphiti-core>=0.3.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.9",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v"
